ENH: rework geogr module
